+++ FAKE/danger.py
@@ -1,0 +1,3 @@
+# infinite.py
+while True:
+    pass