+++ sandbox/samples/hello.py
@@ -1,4 +1,4 @@
 print("Hello Nebula")
 
-with open("artifact.txt","w") as f:
+with open("artifact.txt", "w") as f:
     f.write("Nebula")